Refactor client
