+++ packages/common/src/services/collectibles-provider/CollectiblesProvider.ts
@@ -1,0 +1,5 @@
+import { CollectibleState } from "~/models";
+
+export interface CollectiblesProvider {
+  getCollectibles(wallets: string[]): Promise<CollectibleState>
+}
