+++ packages/common/src/services/collectibles-provider/index.ts
@@ -1,0 +1,2 @@
+export { EthereumCollectiblesProvider } from './EthereumCollectiblesProvider'
+export { SolanaCollectiblesProvider } from './SolanaCollectiblesProvider'
