+++ packages/common/src/services/helius/index.ts
@@ -1,1 +1,2 @@
 export { HeliusClient } from './HeliusClient'
+export * from './types'
