+++ packages/common/src/services/index.ts
@@ -10,7 +10,8 @@ export * from './explore'
 export * from './audio-player'
 export * from './opensea-client'
 export * from './solana-client'
-export * from './helius-client'
+export * from './helius'
+export * from './collectibles-provider'
 export * from './track-download'
 export * from './oauth'
 export * from './location'
