+++ packages/common/src/store/storeContext.ts
@@ -14,15 +14,14 @@ import { AudiusBackend } from '../services/audius-backend'
 import { Env } from '../services/env'
 import { Explore } from '../services/explore'
 import { FingerprintClient } from '../services/fingerprint'
-import { HeliusClient } from '../services/helius-client'
 import { LocalStorage } from '../services/local-storage'
-import { OpenSeaClient } from '../services/opensea-client'
 import { FeatureFlags, RemoteConfigInstance } from '../services/remote-config'
 import { SolanaClient } from '../services/solana-client'
 import { TrackDownload } from '../services/track-download'
 import { WalletClient } from '../services/wallet-client'
 
 import { CommonState } from './reducers'
+import { EthereumCollectiblesProvider, SolanaCollectiblesProvider } from '~/services'
 
 export type CommonStoreContext = {
   getLocalStorageItem: (key: string) => Promise<string | null>
@@ -64,8 +63,9 @@ export type CommonStoreContext = {
     location: Location
   ) => (state: CommonState) => LineupState<Track>
   audioPlayer: AudioPlayer
+  ethereumCollectiblesProvider: EthereumCollectiblesProvider
+  solanaCollectiblesProvider: SolanaCollectiblesProvider
   solanaClient: SolanaClient
-  heliusClient: HeliusClient
   sentry: {
     setTag: (key: string, value: string) => void
     configureScope: (fn: (scope: { setUser: any }) => void) => void
@@ -75,7 +75,6 @@ export type CommonStoreContext = {
   instagramAppId?: string
   instagramRedirectUrl?: string
   share: (url: string, message?: string) => Promise<void> | void
-  openSeaClient: OpenSeaClient
   audiusSdk: () => Promise<AudiusSdk>
   imageUtils: {
     generatePlaylistArtwork: (
