+++ packages/mobile/src/store/storeContext.ts
@@ -1,7 +1,9 @@
 import {
   OpenSeaClient,
   SolanaClient,
-  HeliusClient
+  HeliusClient,
+  SolanaCollectiblesProvider,
+  EthereumCollectiblesProvider
 } from '@audius/common/services'
 import type { CommonStoreContext } from '@audius/common/store'
 import AsyncStorage from '@react-native-async-storage/async-storage'
@@ -42,12 +44,13 @@ export const storeContext: CommonStoreContext = {
   isElectron: false,
   env,
   explore,
-  solanaClient: new SolanaClient({
+  ethereumCollectiblesProvider: new EthereumCollectiblesProvider(new OpenSeaClient(env.OPENSEA_API_URL!)),
+  solanaCollectiblesProvider: new SolanaCollectiblesProvider({
+    heliusClient: new HeliusClient(env.HELIUS_DAS_API_URL),
     solanaClusterEndpoint: env.SOLANA_CLUSTER_ENDPOINT,
     metadataProgramId: env.METADATA_PROGRAM_ID
   }),
-  heliusClient: new HeliusClient({
-    apiUrl: env.HELIUS_DAS_API_URL,
+  solanaClient: new SolanaClient({
     solanaClusterEndpoint: env.SOLANA_CLUSTER_ENDPOINT,
     metadataProgramId: env.METADATA_PROGRAM_ID
   }),
@@ -59,7 +62,6 @@ export const storeContext: CommonStoreContext = {
   instagramAppId: env.INSTAGRAM_APP_ID,
   instagramRedirectUrl: env.INSTAGRAM_REDIRECT_URL,
   share: (url: string, message?: string) => share({ url, message }),
-  openSeaClient: new OpenSeaClient(env.OPENSEA_API_URL!),
   audiusSdk,
   imageUtils: {
     generatePlaylistArtwork
