+++ packages/web/src/common/store/pages/token-dashboard/removeWalletSaga.ts
@@ -13,7 +13,7 @@ import {
 import { call, fork, put, select, takeLatest } from 'typed-redux-saga'
 
 import {
-  fetchOpenSeaNfts,
+  fetchEthereumCollectibles,
   fetchSolanaCollectibles
 } from 'common/store/profile/sagas'
 import { waitForWrite } from 'utils/sagaHelpers'
@@ -121,7 +121,7 @@ function* removeWallet(action: ConfirmRemoveWalletAction) {
     }
 
     yield* fork(fetchSolanaCollectibles, updatedMetadata)
-    yield* fork(fetchOpenSeaNfts, updatedMetadata)
+    yield* fork(fetchEthereumCollectibles, updatedMetadata)
   }
 
   function* onError() {
