+++ packages/web/src/common/store/pages/token-dashboard/sagas.ts
@@ -7,7 +7,7 @@ import {
 import { call, put, select, takeLatest } from 'typed-redux-saga'
 
 import {
-  fetchOpenSeaNftsForWallets,
+  fetchEthereumCollectiblesForWallets,
   fetchSolanaCollectiblesForWallets
 } from 'common/store/profile/sagas'
 import { waitForRead } from 'utils/sagaHelpers'
@@ -27,7 +27,7 @@ function* fetchEthWalletInfo(wallets: string[]) {
   )
 
   const collectiblesMap = (yield* call(
-    fetchOpenSeaNftsForWallets,
+    fetchEthereumCollectiblesForWallets,
     wallets
   )) as CollectibleState
